Sat, Feb 10, 2024, 12:21 AM +01:00

+++ interpreter/interpreter.py
@@ -20,6 +20,6 @@ elif y == '/':
         print("Invalid input. Please enter a valid arithmetic expression.")
         exit()
 
-print(round(result, 1))
+print(float(round(result, 1)))
